+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## v0.2.23 - Transaction Tests
-  - Commit:
+  * Commit: 
   * ResultChecker tests
   * Transaction tests
   * DB interfaces tests
-  - DB adapters tests
+  * DB adapters tests
 ## v0.2.22 - Query Execution Tests
   * Commit: 2026-03-27 11:57
   * ResultChecker functions